made a general slider input
